# njs-backend/src/openApiDocumentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCCMDB",
    "version": "2.2"
  },
  "servers": [
    {
      "url": "http://{servername}/",
      "description": "Node server (dev)",
      "variables": {
        "servername": {
          "description": "name of the server",
          "default": "localhost:8000"
        }
      }
    },
    {
      "url": "https://{servername}/",
      "description": "Node server (production)",
      "variables": {
        "servername": {
          "description": "name of the server",
          "default": "cmdb-backend"
        }
      }
    }
  ],
  "paths": {
...
    "/rest/configurationItem": {
      "post": {
        "summary": "Create a new Configuration item",
        "requestBody": {
          "description": "Data for the new configuration item",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "A configuration item object for creating.",
                "additionalProperties": true,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "pattern": "^.*$",
                    "description": "Name of the configuration item"
                  },
                  "typeId": {
                    "$ref": "#/paths/~1rest~1itemtype~1%7Bid%7D/parameters/0/schema",
                    "description": "Id of the item type"
                  },
                  "attributes": {
                    "type": "array",
                    "description": "Attributes of the item",
                    "items": {
                      "type": "object",
                      "description": "Data for the new attribute",
                      "additionalProperties": false,
                      "properties": {
                        "typeId": {
                          "$ref": "#/paths/~1rest~1itemtype~1%7Bid%7D/parameters/0/schema",
                          "description": "Id of the attribute type, must be allowd by the item type"
                        },
                        "value": {
                          "type": "string",
                          "maxLength": 1000,
                          "pattern": "^.*$",
                          "description": "Attribute value"
                        }
                      }
                    }
                  },
                  "links": {
                    "type": "array",
                    "description": "Links to external websites for the item",
                    "items": {
                      "type": "object",
                      "description": "Data for the new link",
                      "additionalProperties": false,
                      "properties": {
                        "uri": {
                          "type": "string",
                          "maxLength": 1000,
                          "format": "url",
                          "pattern": "^https?://.*$",
                          "description": "Url of the link, must be http or https"
                        },
                        "description": {
                          "$ref": "#/paths/~1rest~1connection~1%7Bid%7D~1description/put/requestBody/content/application~1json/schema/properties/description",
                          "description": "Description for the uri"
                        }
                      }
                    }
                  },
                  "users": {
                    "type": "array",
                    "description": "List of the users that are responsible for the item, i.e. are able to change it",
                    "items": {
                      "$ref": "#/paths/~1login/post/requestBody/content/application~1json/schema/properties/accountName"
                    }
                  }
                },
                "required": [
                  "name",
                  "typeId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationItem~1%7Bid%7D/put/requestBody/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "403": {
            "$ref": "#/paths/~1rest~1user/put/responses/403"
          },
          "415": {
            "$ref": "#/paths/~1login/post/responses/415"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
...
    "/rest/configurationItem/{id}": {
      "parameters": [
        {
          "$ref": "#/paths/~1rest~1itemtype~1%7Bid%7D/parameters/0"
        }
      ],
      "get": {
        "summary": "Get a single configuration item by id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationItem~1%7Bid%7D/put/requestBody/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1rest~1user/put/responses/404"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      },
      "put": {
        "summary": "Update the configuration item",
        "requestBody": {
          "description": "Updated configuration item",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "A configuration item object",
                "allOf": [
                  {
                    "$ref": "#/paths/~1rest~1itemtype~1%7Bid%7D/put/requestBody/content/application~1json/schema/allOf/0"
                  },
                  {
                    "$ref": "#/paths/~1rest~1configurationItem/post/requestBody/content/application~1json/schema"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationItem~1%7Bid%7D/put/requestBody/content/application~1json/schema"
                }
              }
            }
          },
          "304": {
            "description": "nothing changed"
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "403": {
            "$ref": "#/paths/~1rest~1user/put/responses/403"
          },
          "404": {
            "$ref": "#/paths/~1rest~1user/put/responses/404"
          },
          "415": {
            "$ref": "#/paths/~1login/post/responses/415"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      },
      "delete": {
        "summary": "Delete the configuration item type",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationItem~1%7Bid%7D/put/requestBody/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "403": {
            "$ref": "#/paths/~1rest~1user/put/responses/403"
          },
          "404": {
            "$ref": "#/paths/~1rest~1user/put/responses/404"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
...
    "/rest/configurationItem/{id}/full": {
      "parameters": [
        {
          "$ref": "#/paths/~1rest~1itemtype~1%7Bid%7D/parameters/0"
        }
      ],
      "get": {
        "summary": "Get a configuration item with connections by id",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationItem~1full/post/responses/201/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "404": {
            "$ref": "#/paths/~1rest~1user/put/responses/404"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
...
    "/rest/configurationItems": {
      "get": {
        "summary": "Get all configuration items",
        "parameters": [
          {
            "name": "page",
            "allowEmptyValue": true,
            "required": false,
            "description": "Starting page for retrieving items, in portions of 1000 at max",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Returns a portion of the whole number of configuration items.",
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "type": "array",
                      "description": "Array with connection objects belonging to the given item",
                      "maxItems": 1000,
                      "items": {
                        "$ref": "#/paths/~1rest~1configurationItem~1%7Bid%7D/put/requestBody/content/application~1json/schema"
                      }
                    },
                    "totalItems": {
                      "type": "integer",
                      "description": "Number of all items to be found. If it is greater than the count of the items array, then more pages can be retrieved.",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
    "/rest/configurationItems/full": {
      "get": {
        "summary": "Get all configuration items",
        "parameters": [
          {
            "$ref": "#/paths/~1rest~1configurationItems/get/parameters/0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Returns a portion of the whole number of configuration items with connections.",
                  "additionalProperties": false,
                  "properties": {
                    "items": {
                      "type": "array",
                      "description": "Array with connection objects belonging to the given item",
                      "maxLength": 1000,
                      "items": {
                        "$ref": "#/paths/~1rest~1configurationItem~1full/post/responses/201/content/application~1json/schema"
                      }
                    },
                    "totalItems": {
                      "type": "integer",
                      "description": "Number of all items to be found. If it is greater than the count of the items array, then more pages can be retrieved.",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
...
    "/rest/configurationitems/{idList}": {
      "get": {
        "summary": "Get all configurations items with the given ids. If an id cannot be found, no error is thrown.",
        "parameters": [
          {
            "name": "idList",
            "required": true,
            "description": "Single mongo id or comma separated list of ids.",
            "in": "path",
            "schema": {
              "type": "string",
              "maxLength": 1000,
              "pattern": "^[a-f0-9,]*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/paths/~1rest~1configurationitems~1search/post/responses/200/content/application~1json/schema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
    "/rest/configurationitems/{idList}/full": {
      "get": {
        "summary": "Get all configurations items with their connections with the given ids. If an id cannot be found, no error is thrown.",
        "parameters": [
          {
            "$ref": "#/paths/~1rest~1configurationitems~1%7BidList%7D/get/parameters/0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1login/post/responses/400"
          },
          "401": {
            "$ref": "#/paths/~1login/post/responses/401"
          },
          "default": {
            "$ref": "#/paths/~1login/get/responses/default"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  }
}